fix: user update 오류 수정, deleteDto 추가

+++ 52market_back/src/main/java/prgrms/al/back/user/service/UserServiceImpl.java
@@ -3,12 +3,15 @@ package prgrms.al.back.user.service;
 import javassist.NotFoundException;
 import javassist.bytecode.DuplicateMemberException;
 import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import prgrms.al.back.location.domain.Location;
 import prgrms.al.back.location.service.LocationService;
 import prgrms.al.back.user.convertor.UserConvertor;
 import prgrms.al.back.user.domain.User;
+import prgrms.al.back.user.dto.UserDeleteResponse;
 import prgrms.al.back.user.dto.UserDto;
 import prgrms.al.back.user.exception.ErrorCode;
 import prgrms.al.back.user.repository.UserRepository;
@@ -51,8 +54,11 @@ public class UserServiceImpl implements UserService{
 
     @Transactional
     @Override
-    public UserDto updatedUser(UserDto userDto) {
-        User user = repository.findById(userDto.getUserId()).get();
+    public UserDto updatedUser(UserDto userDto) throws NotFoundException {
+        if (!repository.existsById(userDto.getId())){
+            throw new NotFoundException(ErrorCode.NOT_FOUND_USER.toString());
+        }
+        User user = repository.findById(userDto.getId()).get();
         Location location = locationService.findByName(userDto.getLocation()).get();
         user.updateInfo(userDto.getName(), userDto.getNickname(), userDto.getPassword(), location);
         return userConvertor.toDto(repository.findById(user.getId()).get());
@@ -60,8 +66,13 @@ public class UserServiceImpl implements UserService{
 
     @Transactional
     @Override
-    public String deleteUser(Long id) {
+    public UserDeleteResponse deleteUser(Long id) throws NotFoundException{
+
+        if (!repository.existsById(id)){
+            throw new NotFoundException(ErrorCode.NOT_FOUND_USER.toString());
+        }
+
         repository.deleteById(id);
-        return "Deleted User";
+        return UserDeleteResponse.builder().message("success").build();
     }
 }
